% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTransformData.R
\name{Clean_Depth_Data}
\alias{Clean_Depth_Data}
\title{Clean raw user depth data}
\usage{
Clean_Depth_Data(HOBO)
}
\arguments{
\item{HOBO}{The name of the dataframe to be used in processing}
}
\value{
Returns a new dataframe (named "hoboData" by default) containing calculated inundation values based on the user input
}
\description{
Process observed depth data to identify inundation events. Processing includes removing minor fluctuations around the sensor, applying a moving window to smooth high-frequency data (if desired for data sampled <30 min apart), and truncating the record to only observations which occurred pre-hatchling emergence (if desired). A quality control check will prompt the user to continue with or abort processing for records less than 45 days.
}
\examples{
Clean_Depth_Data(HOBO)
}
