% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTransformData.R
\name{Depth_Check}
\alias{Depth_Check}
\title{Check that reported water levels are above logger as opposed to depth below surface}
\usage{
Depth_Check(DATA_Check = Data)
}
\arguments{
\item{DATA_Check}{Uses a processing copy of the inundation data (i.e., "Data") for a QC check}
}
\value{
Water depth above logger in meters, transformed as necessary.
}
\description{
Check that reported water levels are above logger as opposed to depth below surface
}
