% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InspectOutput.R
\name{Export_Tables}
\alias{Export_Tables}
\title{Prepare tables for export to Excel}
\usage{
Export_Tables(Data = inundationStart)
}
\arguments{
\item{Data}{A variable set in the processing phase which determines what tables to prepare for export}
}
\value{
Tables formatted for export to separate Excel sheets
}
\description{
Prepare tables for export to Excel
}
