% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTransformData.R
\name{ID_Inundation}
\alias{ID_Inundation}
\title{Identifies inundation events relative to nest chamber}
\usage{
ID_Inundation()
}
\value{
The start and end Date-Time stamps, duration, and maximum severity of each inundation event
}
\description{
Identifies inundation events relative to nest chamber
}
