% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertPressure.R
\name{kPa_mbar}
\alias{kPa_mbar}
\title{Convert kilopascals to millibars}
\usage{
kPa_mbar(kPa)
}
\arguments{
\item{kPa}{A number in kilopascals}
}
\value{
A number in millibars
}
\description{
Convert kilopascals to millibars
}
