% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InspectOutput.R
\name{Plot_Inspection}
\alias{Plot_Inspection}
\title{Inspect depth and temperature time-series}
\usage{
Plot_Inspection()
}
\value{
A 3-panel plot of cleaned water depth, depth of incursions into nest chamber, and water temperature
}
\description{
Plot water level and temperature for a basic visualization of the inundation time-series. The plot will adapt based on user inputs during Clean_Depth_Data().
}
\examples{
Plot_Inspection()

}
