% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTransformData.R
\name{Process_Normal}
\alias{Process_Normal}
\title{Umbrella function to transform depth data, determine severity, and output results to a new dataset when no moving window is selected}
\usage{
Process_Normal()
}
\value{
A new dataset (i.e., hoboData) containing calculated values
}
\description{
Umbrella function to transform depth data, determine severity, and output results to a new dataset when no moving window is selected
}
