% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InspectOutput.R
\name{Summary_Table}
\alias{Summary_Table}
\title{Summarize inundation exposure across deployment}
\usage{
Summary_Table(HOBO = hoboData)
}
\arguments{
\item{HOBO}{The processed data to be summarized - Defaults to hoboData output by Clean_Depth_Data()}
}
\value{
A table summarizing the length of the observation period, number of inundation events, maximum depth of inundation, various duration factors, and measures of inundation severity
}
\description{
Summarize inundation exposure across deployment
}
