% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InspectOutput.R
\name{Temperature_Table}
\alias{Temperature_Table}
\title{Temperature summary table}
\usage{
Temperature_Table(HOBO = hoboData)
}
\arguments{
\item{HOBO}{The processed data to be summarized - Defaults to hoboData output by Clean_Depth_Data()}
}
\value{
A table summarizing the length of the observation period and descriptive deployment- and daily-level statistics
}
\description{
Generates a table summarizing reported temperature across the logger deployment
}
\examples{
Temperature_Table(HOBO = hoboData)

}
