% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertPressure.R
\name{WaterDepth_kPa}
\alias{WaterDepth_kPa}
\title{Calculate water depth}
\usage{
WaterDepth_kPa(Pressure, Density = 1025)
}
\arguments{
\item{Pressure}{Reported pressure in kilopascals}

\item{Density}{Density of water - preset to 1025 for seawater but can be adjusted by user}
}
\value{
Depth in meters of water above the pressure sensor
}
\description{
Calculate water depth in meters above sensor based on pressure in kilopascals
}
\examples{
WaterDepth_kPa(Pressure = 10)
}
